PlayerSpawnLocation
Added logic to GameStateManager to determine the position of the player similar to how the RoomManager determines it for the enemy.
So with this push, the GameStateManager has a field where a game object with the desired coords is assigned in order to spawn the player at a correct positon.