Checkpoint - Player/Enemy Prefab Update
Changed the structure of the prefabs so that the sprite is now a child of an empty gameobject.
This empty gameobject retains the position of the sprite, while the child PlayerHand/EnemyHand (which both have an animator and a sprite attached) can be animated relative to the empty gameobject's position.

This allows to create animations on the player and enemy hands, without having the animations change the positions of the sprites.