CHECKPOINT - Power up card UX tweak
The power cards can now be hovered over to raise them slightly. There was a bug that cause the cards to not reset back to their initial position after releasing the card, so they were snapping back to their "risen" position, which meant you could keep rising their position with each hover or drag.

TO DO: Right now, the bug is partially fixed, as the cards snap bad to their "risen" position but they reset back to default if you hover over them.