TEST REFACTOR - Flow fixing savepoint
Fixed an issue that caused the room to transition to the next room before defeating all enemies.

Current issue: After defeating both enemies, the game directly transitions to the next room, and only then displays the PowerUpPanel, and if continue is pressed, it takes the player immediately to a next room.